updating folder structure to "best practice"
