+++ .gitignore
@@ -1,4 +1,4 @@
 /.env
-/main
+bin/*
 videos/out/*
 videos/in/*
