+++ internal/controllers/CRUD.go
@@ -8,9 +8,9 @@ import (
 	"github.com/gin-gonic/gin"
 	"gorm.io/gorm"
 	"gorm.io/gorm/clause"
-	"rezeptapp.ml/goApp/error_handler"
-	"rezeptapp.ml/goApp/initializers"
-	"rezeptapp.ml/goApp/models"
+	"github.com/madswillem/recipeApp_Backend_Go/internal/error_handler"
+	"github.com/madswillem/recipeApp_Backend_Go/internal/initializers"
+	"github.com/madswillem/recipeApp_Backend_Go/internal/models"
 )
 
 func GetAll(c *gin.Context) {
@@ -31,39 +31,14 @@ func AddRecipe(c *gin.Context) {
 	err := c.ShouldBindJSON(&body)
 
 	if err != nil {
-		print(err)
 		error_handler.HandleError(c, http.StatusBadRequest, "Failed to read body", err)
 		return
 	}
 
-	err = body.CheckForRequiredFields()
+	err = body.Create()
 	if err != nil {
-		error_handler.HandleError(c, http.StatusBadRequest, "Missing required fields", err)
-		return
-	}
-
-	body.Rating.DefaultRatingStruct(body.Title)
-	for i := 0; i < len(body.Ingredients); i++ {
-		body.Ingredients[i].Rating.DefaultRatingStruct(body.Ingredients[i].Ingredient)
-	}
-
-	err = body.AddNutritionalValue()
-	if err != nil {
-		error_handler.HandleError(c, http.StatusBadRequest, "Database error", err)
-		return
-	}
-
-	tx := initializers.DB.Begin()
-
-	if err := tx.Create(&body).Error; err != nil {
-		tx.Rollback()
-		error_handler.HandleError(c, http.StatusBadRequest, "Database error", err)
-		return
-	}
-
-	if err := tx.Commit().Error; err != nil {
-		error_handler.HandleError(c, http.StatusBadRequest, "Database error", err)
-		return
+		error_handler.HandleError(c, http.StatusBadRequest, "Failed to create recipe", err)
+		return	
 	}
 
 	c.JSON(http.StatusCreated, body)
@@ -77,9 +52,9 @@ func UpdateRecipe(c *gin.Context) {
 	}
 
 	var body models.RecipeSchema
-	body.ID = uint(i)
-
 	c.ShouldBindJSON(&body)
+
+	body.ID = uint(i)
 	
 	err = body.Update(c)
 	if err != nil {
@@ -119,7 +94,15 @@ func GetById(c *gin.Context) {
 		return
 	}
 	response := models.RecipeSchema{ID: uint(i)}
-	err = response.GetRecipeByID(c)
+	reqData := map[string]bool{
+		"ingredients":      true,
+		"ingredient_nutri": true,
+		"ingredient_rate":  true,
+		"rating":           true,
+		"nutritionalvalue": true,
+		"diet":             true,
+	}
+	err = response.GetRecipeByID(c, reqData)
 
 	if err != nil {
 		if err == gorm.ErrRecordNotFound {
