+++ internal/controllers/render.go
@@ -7,9 +7,9 @@ import (
 	"github.com/gin-gonic/gin"
 	"gorm.io/gorm"
 	"gorm.io/gorm/clause"
-	"rezeptapp.ml/goApp/error_handler"
-	"rezeptapp.ml/goApp/initializers"
-	"rezeptapp.ml/goApp/models"
+	"github.com/madswillem/recipeApp_Backend_Go/internal/error_handler"
+	"github.com/madswillem/recipeApp_Backend_Go/internal/initializers"
+	"github.com/madswillem/recipeApp_Backend_Go/internal/models"
 )
 
 func RenderHome(c *gin.Context) {
@@ -40,7 +40,15 @@ func RenderProductpage(c *gin.Context) {
 	}
 	
 	res := models.RecipeSchema{ID: uint(i)}
-	err = res.GetRecipeByID(c)
+	reqData := map[string]bool{
+		"ingredients":      true,
+		"ingredient_nutri": true,
+		"ingredient_rate":  true,
+		"rating":           true,
+		"nutritionalvalue": true,
+		"diet":             true,
+	}
+	err = res.GetRecipeByID(c, reqData)
 
 	if err != nil {
 		if err == gorm.ErrRecordNotFound {
