+++ internal/error_handler/handleError.go
@@ -3,7 +3,7 @@ package error_handler
 import "github.com/gin-gonic/gin"
 func HandleError(c *gin.Context, statusCode int, errorMessage string, err error) {
 	c.AbortWithStatusJSON(statusCode, gin.H{
-		"error":      errorMessage,
-		"errMessage": err.Error(),
+		"error":      err,
+		"errMessage": errorMessage,
 	})
 }
