+++ internal/models/dietModel.go
@@ -1,16 +1,5 @@
 package models
 
-type IngredientsSchema struct {
-	ID        		 uint 			 	 `json:"id" gorm:"primarykey"`
-	RecipeSchemaID 	 uint				 `json:"-"`
-
-	Ingredient		 string				 `json:"ingredient"` 
-	Amount			 string				 `json:"amount"` 
-	MeasurementUnit	 string				 `json:"measurement_unit"`
-	NutritionalValue NutritionalValue    `json:"nutritional_value" gorm:"polymorphic:Owner"`
-	Rating			 RatingStruct		 `json:"rating" gorm:"polymorphic:Owner"`
-}
-
 type DietSchema struct {
 	ID		  		 uint 			 	 `gorm:"primarykey"`
 	OwnerTitle	string 					 `json:"owner_title"`
@@ -27,3 +16,4 @@ type DietSchema struct {
 	FoodCombining	 bool				 `json:"food_combining"`
 	WholeFood		 bool				 `json:"whole_food"`
 }
+
