+++ internal/models/recipeModel.go
@@ -2,15 +2,12 @@ package models
 
 import (
 	"errors"
-	"fmt"
 	"net/http"
 
 	"github.com/gin-gonic/gin"
 	"gorm.io/gorm"
-	"gorm.io/gorm/clause"
-	"rezeptapp.ml/goApp/error_handler"
-	"rezeptapp.ml/goApp/initializers"
-	"rezeptapp.ml/goApp/tools"
+	"github.com/madswillem/recipeApp_Backend_Go/internal/error_handler"
+	"github.com/madswillem/recipeApp_Backend_Go/internal/initializers"
 )
 
 type RecipeSchema struct {
@@ -62,8 +59,27 @@ func (recipe *RecipeSchema) CheckIfExistsByID() (bool, error) {
 	return result.Found, err
 }
 
-func (recipe *RecipeSchema) GetRecipeByID(c *gin.Context) error {
-	err := initializers.DB.Preload(clause.Associations).Preload("Ingredients.NutritionalValue").Preload("Ingredients.Rating").Preload("Ingredients.NutritionalValue").First(&recipe, "ID = ?", recipe.ID).Error
+func (recipe *RecipeSchema) GetRecipeByID(c *gin.Context, reqData map[string]bool) error {
+	req := initializers.DB
+	if reqData["ingredients"] || reqData["everything"] {
+		req = req.Preload("Ingredients")
+	}
+	if reqData["ingredient_nutri"] || reqData["everything"] {
+		req = req.Preload("Ingredients.NutritionalValue")
+	}
+	if reqData["ingredient_rate"] || reqData["everything"] {
+		req = req.Preload("Ingredients.Rating")
+	}
+	if reqData["rating"] || reqData["everything"] {
+		req = req.Preload("Rating")
+	}
+	if reqData["nutritionalvalue"] || reqData["everything"] {
+		req = req.Preload("NutritionalValue")
+	}
+	if reqData["diet"] || reqData["everything"] {
+		req = req.Preload("Diet")
+	}
+	err := req.First(&recipe, "ID = ?", recipe.ID).Error
 
 	if err != nil {
 		if err == gorm.ErrRecordNotFound {
@@ -146,115 +162,32 @@ func (recipe *RecipeSchema) CheckForRequiredFields() error {
 	return nil
 }
 
-func (rating RatingStruct) Update(change int, c *gin.Context) (RatingStruct, error) {
-
-	result := rating
-	data, err := tools.GetCurrentData()
-
-	if err != nil {
-		return result, err
-	}
-
-	percentage := 10.0
-
-	switch data.Day {
-	case "Mon":
-		result.Mon += tools.PercentageCalculator(result.Mon*float64(change), percentage)
-	case "Tue":
-		result.Tue += tools.PercentageCalculator(result.Tue*float64(change), percentage)
-	case "Wed":
-		result.Wed += tools.PercentageCalculator(result.Wed*float64(change), percentage)
-	case "Thu":
-		result.Thu += tools.PercentageCalculator(result.Thu*float64(change), percentage)
-	case "Fri":
-		result.Fri += tools.PercentageCalculator(result.Fri*float64(change), percentage)
-	case "Sat":
-		result.Sat += tools.PercentageCalculator(result.Sat*float64(change), percentage)
-	case "Sun":
-		result.Sun += tools.PercentageCalculator(result.Sun*float64(change), percentage)
-	default:
-		c.AbortWithStatus(http.StatusInternalServerError)
-	}
-
-	switch data.Season {
-	case "Win":
-		result.Win += tools.PercentageCalculator(result.Win*float64(change), percentage)
-	case "Spr":
-		result.Spr += tools.PercentageCalculator(result.Spr*float64(change), percentage)
-	case "Sum":
-		result.Sum += tools.PercentageCalculator(result.Sum*float64(change), percentage)
-	case "Aut":
-		result.Aut += tools.PercentageCalculator(result.Aut*float64(change), percentage)
-	default:
-		c.AbortWithStatus(http.StatusInternalServerError)
-	}
-
-	switch data.Temp {
-	case "subzerodegree":
-		result.Subzerodegree += tools.PercentageCalculator(result.Subzerodegree*float64(change), percentage)
-	case "zerodegree":
-		result.Zerodegree += tools.PercentageCalculator(result.Zerodegree*float64(change), percentage)
-	case "tendegree":
-		result.Tendegree += tools.PercentageCalculator(result.Tendegree*float64(change), percentage)
-	case "twentiedegree":
-		result.Twentiedegree += tools.PercentageCalculator(result.Twentiedegree*float64(change), percentage)
-	case "thirtydegree":
-		result.Thirtydegree += tools.PercentageCalculator(result.Thirtydegree*float64(change), percentage)
-	}
-
-	arr := []float64{
-		result.Mon,
-		result.Tue,
-		result.Wed,
-		result.Thu,
-		result.Fri,
-		result.Sat,
-		result.Sun,
-
-		result.Win,
-		result.Spr,
-		result.Sum,
-		result.Aut,
-
-		result.Thirtydegree,
-		result.Twentiedegree,
-		result.Tendegree,
-		result.Zerodegree,
-		result.Subzerodegree,
-	};
-
-	result.Overall = tools.CalculateAverage(arr)
-	fmt.Println(result.Overall)
-
-	return result, err
-}
-
-
-
-func (ingredient *IngredientsSchema) createIngredientDBEntry() error {
-	newIngredientDBEntry := IngredientDBSchema{
-		Name:       ingredient.Ingredient,
-		StandardUnit: ingredient.MeasurementUnit,
-		NutritionalValue: ingredient.NutritionalValue,
-	}
-	
-	err := initializers.DB.Create(&newIngredientDBEntry).Error
+func (recipe *RecipeSchema) Create() error {
+	err := recipe.CheckForRequiredFields()
 	if err != nil {
 		return err
 	}
 
-	return err
-}
+	recipe.Rating.DefaultRatingStruct(recipe.Title)
+	for i := 0; i < len(recipe.Ingredients); i++ {
+		recipe.Ingredients[i].Rating.DefaultRatingStruct(recipe.Ingredients[i].Ingredient)
+	}
 
-func (ingredient *IngredientsSchema) CheckForRequiredFields() error {
-	if ingredient.Ingredient == "" {
-		return errors.New("missing ingredient")
+	err = recipe.AddNutritionalValue()
+	if err != nil {
+		return err
 	}
-	if ingredient.Amount == "" {
-		return errors.New("missing amount")
+
+	tx := initializers.DB.Begin()
+
+	if err := tx.Create(&recipe).Error; err != nil {
+		tx.Rollback()
+		return err
 	}
-	if ingredient.MeasurementUnit == "" {
-		return errors.New("missing measurement unit")
+
+	if err := tx.Commit().Error; err != nil {
+		return err
 	}
-	return nil
+
+	return err
 }
