+++ internal/tools/math.go
