+++ test/testdata/create/add_recipe_with_all_required_fields(edited=false).json
